backend/backend/requirements.txt
Flask==3.0.0
Flask-SocketIO==5.3.6
python-dotenv==1.0.0
PyJWT==2.8.0
python-engineio==4.9.0
python-socketio==5.10.0
requests==2.31.0
openai==1.3.7
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Werkzeug==3.0.1
bcrypt==4.0.1
python-dateutil==2.8.2
